+++ bitreader/include/bitreader/common/file_reader.hpp
@@ -8,7 +8,7 @@ namespace brcpp
     class file_reader
     {
     public:
-        virtual size_t read(uint8_t* dest, uint64_t position, size_t bytes) = 0;
+        virtual size_t read(uint8_t* dest, size_t position, size_t bytes) = 0;
         virtual size_t size() = 0;
         virtual bool depleted() = 0;
         virtual std::shared_ptr<file_reader> clone() = 0;
